Add command support for Windows systems and drop no-pty shell
This commit avoids code duplication in ssh_session_*.go by introducing
ssh_session.go and removes support for the awkwared non-pty shell for
windows versions prior to Win10 Build 17763 without ssh-shellhost.exe.

Better shell access can now be achieved via command execution.
