+++ ssh_session_windows.go
@@ -19,7 +19,6 @@
 package main
 
 import (
-	"bufio"
 	"io"
 	"log"
 	"os"
@@ -31,128 +30,87 @@ import (
 	"golang.org/x/sys/windows"
 )
 
-func makeSSHSessionHandler(shell string) ssh.Handler {
-	return func(s ssh.Session) {
-		log.Printf("New login from %s@%s", s.User(), s.RemoteAddr().String())
-		ptyReq, winCh, isPty := s.Pty()
-		if isPty {
-			vsn := windows.RtlGetVersion()
-			if vsn.MajorVersion < 10 ||
-				vsn.BuildNumber < 17763 {
-				// Non-interactive Pty
+func createPty(s ssh.Session, shell string) {
+	ptyReq, winCh, _ := s.Pty()
+	vsn := windows.RtlGetVersion()
+	if vsn.MajorVersion < 10 ||
+		vsn.BuildNumber < 17763 {
+		// Interactive Pty via ssh-shellhost.exe
 
-				log.Println("Windows version too old to support ConPTY shell")
+		log.Println("Windows version too old to support ConPTY shell")
 
-				var cmd *exec.Cmd
-				if shell == defaultShell {
-					log.Println("Launching primitive cmd hook")
-
-					cmd = exec.Command("cmd")
-					cmd.Stdout = s
-					cmd.Stderr = s
-					stdin, err := cmd.StdinPipe()
-					if err != nil {
-						log.Println("Could not create StdInPipe:", err)
-						s.Exit(1)
-						return
-					}
-
-					// Handle input from client
-					go func() {
-						for {
-							// Echo incoming chars back to client
-							echo := io.TeeReader(s, s)
-
-							// Read until next unix line ending (\n)
-							data, err := bufio.NewReader(echo).ReadBytes(byte(13))
-							if err == io.EOF {
-								log.Println("Connection closed by client")
-								return
-							} else if err != nil {
-								log.Println("Error while reading from client:", err)
-							}
-							if len(data) > 0 {
-								// Send linebreak to client for readability
-								s.Write([]byte("\n"))
-
-								// Write data to stdin of cmd with appended windows line endings
-								stdin.Write(append(data[:len(data)-1], "\r\n"...))
-							}
-						}
-					}()
-				} else {
-					log.Println("Launching shell with ssh-shellhost.exe")
-
-					cmd = exec.Command(shell)
-					cmd.SysProcAttr = &syscall.SysProcAttr{
-						HideWindow:    true,
-						CmdLine:       " " + "---pty cmd", // Must leave a space to the beginning
-						CreationFlags: 0x08000000,
-					}
-					cmd.Stdout = s
-					cmd.Stderr = s
-					cmd.Stdin = s
-
-				}
-
-				if err := cmd.Run(); err != nil {
-					log.Println("Session ended with error:", err)
-					s.Exit(1)
-				} else {
-					log.Println("Session ended normally")
-					s.Exit(0)
-				}
-
-			} else {
-				// Interactive ConPTY
-
-				cpty, err := conpty.New(int16(ptyReq.Window.Width), int16(ptyReq.Window.Height))
-				if err != nil {
-					log.Fatalf("Could not open a conpty terminal: %v", err)
-				}
-				defer cpty.Close()
-
-				// Dynamically handle resizes of terminal window
-				go func() {
-					for win := range winCh {
-						cpty.Resize(uint16(win.Width), uint16(win.Height))
-					}
-				}()
-
-				// Spawn and catch new powershell process
-				pid, _, err := cpty.Spawn(
-					"C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
-					[]string{},
-					&syscall.ProcAttr{
-						Env: os.Environ(),
-					},
-				)
-				if err != nil {
-					log.Fatalf("Could not spawn a powershell: %v", err)
-				}
-				log.Printf("New process with pid %d spawned", pid)
-				process, err := os.FindProcess(pid)
-				if err != nil {
-					log.Fatalf("Failed to find process: %v", err)
-				}
-
-				// Link data streams of ssh session and conpty
-				go io.Copy(s, cpty.OutPipe())
-				go io.Copy(cpty.InPipe(), s)
-
-				ps, err := process.Wait()
-				if err != nil {
-					log.Printf("Error waiting for process: %v", err)
-					s.Exit(1)
-					return
-				}
-				log.Printf("Session ended normally, exit code %d", ps.ExitCode())
-				s.Exit(ps.ExitCode())
-			}
-
-		} else {
-			io.WriteString(s, "Remote forwarding available...\n")
-			select {}
+		if shell == defaultShell {
+			log.Println("No fully interactive shell available, denying PTY request")
+			io.WriteString(s, "No ConPTY shell or ssh-shellhost enhanced shell available. "+
+				"Please append 'cmd' to your ssh command to gain shell access, i.e. "+
+				"'ssh <OPTIONS> <IP> cmd'.\n")
+			s.Exit(1)
+			return
 		}
+		log.Println("Launching shell with ssh-shellhost.exe")
+
+		cmd := exec.Command(shell)
+		cmd.SysProcAttr = &syscall.SysProcAttr{
+			HideWindow:    true,
+			CmdLine:       " " + "---pty cmd", // Must leave a space to the beginning
+			CreationFlags: 0x08000000,
+		}
+		cmd.Stdout = s
+		cmd.Stderr = s
+		cmd.Stdin = s
+
+		if err := cmd.Run(); err != nil {
+			log.Println("Session ended with error:", err)
+			s.Exit(1)
+		} else {
+			log.Println("Session ended normally")
+			s.Exit(0)
+		}
+
+	} else {
+		// Interactive ConPTY
+
+		cpty, err := conpty.New(int16(ptyReq.Window.Width), int16(ptyReq.Window.Height))
+		if err != nil {
+			log.Fatalf("Could not open a conpty terminal: %v", err)
+		}
+		defer cpty.Close()
+
+		// Dynamically handle resizes of terminal window
+		go func() {
+			for win := range winCh {
+				cpty.Resize(uint16(win.Width), uint16(win.Height))
+			}
+		}()
+
+		// Spawn and catch new powershell process
+		pid, _, err := cpty.Spawn(
+			"C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\powershell.exe",
+			[]string{},
+			&syscall.ProcAttr{
+				Env: os.Environ(),
+			},
+		)
+		if err != nil {
+			log.Fatalf("Could not spawn a powershell: %v", err)
+		}
+		log.Printf("New process with pid %d spawned", pid)
+		process, err := os.FindProcess(pid)
+		if err != nil {
+			log.Fatalf("Failed to find process: %v", err)
+		}
+
+		// Link data streams of ssh session and conpty
+		go io.Copy(s, cpty.OutPipe())
+		go io.Copy(cpty.InPipe(), s)
+
+		ps, err := process.Wait()
+		if err != nil {
+			log.Printf("Error waiting for process: %v", err)
+			s.Exit(1)
+			return
+		}
+		log.Printf("Session ended normally, exit code %d", ps.ExitCode())
+		s.Exit(ps.ExitCode())
 	}
 }
